requirements.txt
asgiref==3.5.2
astroid==2.12.12
colorama==0.4.6
dill==0.3.6
dj-database-url==1.0.0
Django==4.1.3
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
Faker==15.3.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
pkg==0.2
platformdirs==2.5.3
psycopg2==2.9.5
pylint==2.15.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
tzdata==2022.6
validate-docbr==1.10.0
whitenoise==6.2.0
wrapt==1.14.1
Pillow==9.1.0
